[UTIL] Correct param type for encoder

+++ runtime_util/runtime_util.c
@@ -89,7 +89,7 @@ device_t KoalaBear = {
         {.name = "pid_kp_a", .type = FLOAT, .read = 1, .write = 1},
         {.name = "pid_ki_a", .type = FLOAT, .read = 1, .write = 1},
         {.name = "pid_kd_a", .type = FLOAT, .read = 1, .write = 1},
-        {.name = "enc_a", .type = FLOAT, .read = 1, .write = 1},
+        {.name = "enc_a", .type = INT, .read = 1, .write = 1},
         // Same params as above except for motor b
         {.name = "duty_cycle_b", .type = FLOAT, .read = 1, .write = 1},
         {.name = "deadband_b", .type = FLOAT, .read = 1, .write = 1},
@@ -98,7 +98,7 @@ device_t KoalaBear = {
         {.name = "pid_kp_b", .type = FLOAT, .read = 1, .write = 1},
         {.name = "pid_ki_b", .type = FLOAT, .read = 1, .write = 1},
         {.name = "pid_kd_b", .type = FLOAT, .read = 1, .write = 1},
-        {.name = "enc_b", .type = FLOAT, .read = 1, .write = 1},
+        {.name = "enc_b", .type = INT, .read = 1, .write = 1},
     }};
 
 // *********************** VIRTUAL DEVICE DEFINITIONS *********************** //
